+++ previous-versions/common/RELEASES.md
@@ -601,3 +601,13 @@ Based on:
 - [python v0.1.0] previous-versions/common
 ### Releases
 - [PyPI v0.1.0] https://pypi.org/project/codat-common/0.1.0 - previous-versions/common
+
+## 2023-09-07 00:15:47
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Common.yaml
+- Speakeasy CLI 1.79.0 (2.101.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.1.0] previous-versions/common
+### Releases
+- [PyPI v0.1.0] https://pypi.org/project/codat-common/0.1.0 - previous-versions/common